Code tidying.

+++ include/serial.h
@@ -29,7 +29,7 @@ public:
     void writeInt(const int_like val)
     {
         char out_str[255];
-        const int len = string_from_int(val, out_str);
+        const size_t len = string_from_int(val, out_str);
         char trimmed_str[len];
         for (size_t j = 0; j < len; j++)
         {
